@@ -17,7 +17,7 @@ prometheus-fastapi-instrumentator = "^6.1.0"
 psycopg2-binary = "^2.9.8"
 pydantic = {extras = ["email"], version = "^2.4.2"}
 pydantic-settings = "^2.0.3"
-python = "^3.11"
+python = "^3.12"
 python-dotenv = "^1.0.0"
 python-jose = "^3.3.0"
 python-multipart = "^0.0.6"
@@ -27,7 +27,7 @@ SQLAlchemy = "^2.0.21"
 tenacity = "^8.2.3"
 types-redis = "^4.6.0.7"
 uvicorn = "^0.23.2"
-uvloop = "^0.18.0"
+uvloop = "^0.19.0"
 
 [tool.poetry.group.dev.dependencies]
 alembic = "^1.12.0"
