@@ -1,6 +1,7 @@
+from typing import Dict
+
 from requests import Response, Session
 from requests.adapters import HTTPAdapter
-from typing import Dict
 from urllib3 import disable_warnings
 from urllib3.exceptions import InsecureRequestWarning
 
