@@ -7,7 +7,7 @@ from {{cookiecutter.package_name}}.exceptions import (
     InternalServerError,
     NotFoundError,
     UnauthenticatedError,
-    ValidationError
+    ValidationError,
 )
 
 
