@@ -35,7 +35,7 @@ def test_get_logger_config(mocker):
     settings_mock = MagicMock(spec=Settings)
     settings_mock.HOSTNAME = "hostname"
     settings_mock.LOG_FILE_DIRECTORY = "log-file-directory"
-    settings_mock.LOG_FILE_MAX_BYTES =1000
+    settings_mock.LOG_FILE_MAX_BYTES = 1000
     settings_mock.LOG_FILE_BACKUP_COUNT = 1
     mocker.patch.object(config, "settings", settings_mock)
 
