@@ -4,5 +4,5 @@ from .http import (
     InternalServerError,
     NotFoundError,
     UnauthenticatedError,
-    ValidationError
+    ValidationError,
 )
